KAFKA-8091; Wait for processor shutdown before testing removed listeners (#6425)
DynamicBrokerReconfigurationTest.testAddRemoveSaslListeners removes a listener, waits for the config to be propagated to all brokers and then validates that connections to the removed listener fail. But there is a small timing window between config update and Processor shutdown. Before validating that connections to a removed listener fail, this commit waits for all metrics of the removed listener to be deleted, ensuring that the Processors of the listener have been shutdown.

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>