KAFKA-7813: JmxTool throws NPE when --object-name is omitted
https://issues.apache.org/jira/browse/KAFKA-7813

Running the JMX tool without --object-name parameter, results in a NullPointerException.

*More detailed description of your change,
if necessary. The PR title and PR message become
the squashed commit message, so use a separate
comment to ping reviewers.*

*Summary of testing strategy (including rationale)
for the feature or bug fix. Unit and/or integration
tests are expected for any behaviour change and
system tests should be considered for larger changes.*

Author: huxihx <huxi_2b@hotmail.com>

Reviewers: Ewen Cheslack-Postava <ewen@confluent.io>

Closes #6139 from huxihx/KAFKA-7813